requirements.txt
marshmallow==3.12.1
marshmallow-enum==1.5.1
matplotlib==3.3.2
Pillow==7.2.0
jupyter